*Fixed equations in EKF.updateFromReobservedFeatures()
(1) Fixed the computation for the lower left matrix. The formula was
lacking a final transpose operation after every operation was done.
(2) Moved the code for addition of columns based on the transposition of
the last rows to after the last rows have been initialized. Also looped
only until toAdd.size()-1 because the last entry (lower right matrix)
shouldn't be transposed anymore.